Fix erro

+++ app/src/main/java/com/example/money_manager_app/data/dao/DebtTransactionDao.kt
@@ -48,7 +48,7 @@ interface DebtTransactionDao {
         endDate : Long?,
         minAmount : Double?,
         maxAmount : Double?,
-        categoryType: Int?,
+        categoryType: Long?,
         fromWallet : Long?,
         idAccount : Long
     ): List<DebtTransaction>
